requirements.txt

biopython==1.72
h5py==2.8.0
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
numpy==1.14.5
pandas==0.23.4

scikit-learn==0.19.2
scipy==1.1.0
tensorflow-gpu==1.10.0
jupyter==1.0.0
seaborn==0.11.1
seqlogo==5.29.8
orange3==3.26.0
